+++ libraries/base/Foreign/Marshal/Utils.hs
@@ -43,13 +43,18 @@ module Foreign.Marshal.Utils (
   --
   copyBytes,
   moveBytes,
+
+  -- ** Filling up memory area with required values
+  --
+  fillBytes,
 ) where
 
 import Data.Maybe
 import Foreign.Ptr              ( Ptr, nullPtr )
 import Foreign.Storable         ( Storable(poke) )
-import Foreign.C.Types          ( CSize(..) )
+import Foreign.C.Types          ( CSize(..), CInt(..) )
 import Foreign.Marshal.Alloc    ( malloc, alloca )
+import Data.Word                ( Word8 )
 
 import GHC.Real                 ( fromIntegral )
 import GHC.Num
@@ -161,6 +166,16 @@ moveBytes               :: Ptr a -> Ptr a -> Int -> IO ()
 moveBytes dest src size  = do _ <- memmove dest src (fromIntegral size)
                               return ()
 
+-- Filling up memory area with required values
+-- -------------------------------------------
+
+-- |Fill a given number of bytes in memory area with a byte value.
+--
+-- /Since: 4.8.0.0/
+fillBytes               :: Ptr a -> Word8 -> Int -> IO ()
+fillBytes dest char size = do
+  _ <- memset dest (fromIntegral char) (fromIntegral size)
+  return ()
 
 -- auxilliary routines
 -- -------------------
@@ -169,4 +184,4 @@ moveBytes dest src size  = do _ <- memmove dest src (fromIntegral size)
 --
 foreign import ccall unsafe "string.h" memcpy  :: Ptr a -> Ptr a -> CSize -> IO (Ptr a)
 foreign import ccall unsafe "string.h" memmove :: Ptr a -> Ptr a -> CSize -> IO (Ptr a)
-
+foreign import ccall unsafe "string.h" memset  :: Ptr a -> CInt  -> CSize -> IO (Ptr a)